Renamed interfaces and moved things around
