+++ example/src/main/java/uk/co/kiwisoft/validroid_examples/MyHandler.java
@@ -1,11 +1,11 @@
 package uk.co.kiwisoft.validroid_examples;
 
-import uk.co.kiwisoft.validroid.handlers.IHandler;
+import uk.co.kiwisoft.validroid.Handler;
 
 /**
  * ToDo: Explain the scope of the class
  */
-public class MyHandler implements IHandler<MyCustomView> {
+public class MyHandler implements Handler<MyCustomView> {
 
     public void handleErrorMessages(String[] errors) {
         for (String msg : errors) {
