+++ example/src/main/java/uk/co/kiwisoft/validroid_examples/MyProvider.java
@@ -1,12 +1,12 @@
 package uk.co.kiwisoft.validroid_examples;
 
 
-import uk.co.kiwisoft.validroid.providers.IProvider;
+import uk.co.kiwisoft.validroid.Provider;
 
 /**
  * ToDo: Explain the scope of the class
  */
-public class MyProvider implements IProvider<String, MyCustomView> {
+public class MyProvider implements Provider<String, MyCustomView> {
 
     public String provideData(MyCustomView container) {
         return container.getMyData();
