+++ example/src/main/java/uk/co/kiwisoft/validroid_examples/validators/EmailValidator.java
@@ -3,12 +3,12 @@ package uk.co.kiwisoft.validroid_examples.validators;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import uk.co.kiwisoft.validroid.validators.IValidator;
+import uk.co.kiwisoft.validroid.Validator;
 
 /**
  * ToDo: Explain the scope of the class
  */
-public class EmailValidator implements IValidator<String> {
+public class EmailValidator implements Validator<String> {
     private static final Pattern EMAIL_PATTERN =
             Pattern.compile("^[A-Z0-9._%+-]+@[A-Z0-9.-]+\\.[A-Z]{2,6}$", Pattern.CASE_INSENSITIVE);
