+++ example/src/main/java/uk/co/kiwisoft/validroid_examples/validators/LengthValidator.java
@@ -1,11 +1,11 @@
 package uk.co.kiwisoft.validroid_examples.validators;
 
-import uk.co.kiwisoft.validroid.validators.IValidator;
+import uk.co.kiwisoft.validroid.Validator;
 
 /**
  * ToDo: Explain the scope of the class
  */
-public class LengthValidator implements IValidator<String> {
+public class LengthValidator implements Validator<String> {
     public boolean isValid(String input) {
         return input.length() >= 6 && input.length() <= 50;
     }
