+++ validroid/src/main/java/uk/co/kiwisoft/validroid/Handler.java
@@ -1,8 +1,8 @@
-package uk.co.kiwisoft.validroid.handlers;
+package uk.co.kiwisoft.validroid;
 
 /**
  * ToDo: Explain the scope of the class
  */
-public interface IHandler<O> {
+public interface Handler {
     void handleErrorMessages(String[] errors);
 }
