+++ validroid/src/main/java/uk/co/kiwisoft/validroid/ValidationRunner.java
@@ -5,7 +5,7 @@ import java.util.List;
 /**
  * Interface to create validation runners
  */
-public interface IValidationRunner {
+public interface ValidationRunner {
     void addWork(WorkContainer workContainer);
 
     void addWork(List<WorkContainer> workContainers);
