+++ validroid/src/main/java/uk/co/kiwisoft/validroid/Validator.java
@@ -1,9 +1,9 @@
-package uk.co.kiwisoft.validroid.validators;
+package uk.co.kiwisoft.validroid;
 
 /**
  * ToDo: Explain the scope of the class
  */
-public interface IValidator<T> {
+public interface Validator<T> {
     boolean isValid(T input);
     String[] getErrorMessages();
 }
